backend-lab/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
minio==7.2.3
aiofiles==23.2.1
httpx==0.26.0
redis==5.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
pillow==10.2.0
python-magic==0.4.27
boto3==1.34.14
celery==5.3.6
pydicom==2.4.3
reportlab==4.0.8
structlog==24.1.0
tenacity==8.2.3
